requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
decorator==4.4.2
fonttools==4.29.1
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
kiwisolver==1.3.2
matplotlib==3.5.1
moviepy==1.0.3
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
proglog==0.1.9
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
tqdm==4.62.3
urllib3==1.26.8
